core, graph, graphql, server: Allow to override the `first` limit
This is useful for internal queries where we want _all_ the results,
not just the first 100. We could paginate in those cases but it'll
only mean more store queries to be executed.

+++ graph/src/data/query/error.rs
@@ -36,7 +36,7 @@ pub enum QueryExecutionError {
     EmptyQuery,
     MultipleSubscriptionFields,
     SubgraphDeploymentIdError(String),
-    RangeArgumentsError(Vec<&'static str>, u64),
+    RangeArgumentsError(Vec<&'static str>, u32),
     InvalidFilterError,
     EntityFieldError(String, String),
     ListTypesError(String, Vec<String>),
